Recover four USA functions and reach 40.00 percent
Forty-two single-function workers across three populations. The per-stream
attribution matters more than the total, so it is recorded here rather than
inferred from the coverage jump:

- near-misses at medium effort: 14 assigned, 2 exact, 12 still correct size
- wrong-size at low effort:     14 assigned, 2 exact
- uncovered at low effort:      14 assigned, 0 exact, 0 promotions

Raising the near-misses from low to medium moved that population from 0 of 14
to 2 of 14, so the previous wave's null result was a resourcing problem rather
than a stuck population. But the wrong-size stream reached the same rate at low
effort, which is the cheaper way to buy a match, and the uncovered stream
returned neither a match nor a promotion. Its remaining members average about
124 bytes and are largely label-scan artifacts rather than functions.

Token cost is recorded per function for all forty-two workers in
docs/function_token_log.tsv, each measured from that worker's own transcript.

+++ src/functions/func_8006FB00.c
@@ -65,8 +65,8 @@ void func_8006FB00(Unk8006FB00 *arg0) {
     if (arg0->unk60 & 0x80) {
         arg0->unk18C = arg0->unk18C * arg0->unk1AC;
     }
-    if (0.0f < arg0->unk264) {
-        arg0->unk18C = 0.0f;
+    if ((f32)0 < arg0->unk264) {
+        arg0->unk18C = (f32)0;
     }
     scale = arg0->unk22C;
     limit = arg0->unk74 * scale;
